move SQLViewSuite to sql/core.
