+++ sql/core/src/test/scala/org/apache/spark/sql/execution/SQLViewSuite.scala
@@ -15,30 +15,35 @@
  * limitations under the License.
  */
 
-package org.apache.spark.sql.hive.execution
+package org.apache.spark.sql.execution
 
-import org.apache.spark.sql.{AnalysisException, QueryTest, Row, SaveMode}
+import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.NoSuchTableException
-import org.apache.spark.sql.catalyst.catalog.{CatalogStorageFormat, CatalogTable, CatalogTableType}
-import org.apache.spark.sql.hive.test.TestHiveSingleton
-import org.apache.spark.sql.test.SQLTestUtils
-import org.apache.spark.sql.types.StructType
+import org.apache.spark.sql.test.{SharedSQLContext, SQLTestUtils}
 
-/**
- * A suite for testing view related functionality.
- */
-class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
-  import spark.implicits._
+class SimpleSQLViewSuite extends SQLViewSuite with SharedSQLContext {
 
   override def beforeAll(): Unit = {
+    super.beforeAll()
     // Create a simple table with two columns: id and id1
     spark.range(1, 10).selectExpr("id", "id id1").write.format("json").saveAsTable("jt")
   }
 
   override def afterAll(): Unit = {
-    spark.sql(s"DROP TABLE IF EXISTS jt")
+    try {
+      spark.sql(s"DROP TABLE IF EXISTS jt")
+    } finally {
+      super.afterAll()
+    }
   }
+}
+
+/**
+ * A suite for testing view related functionality.
+ */
+abstract class SQLViewSuite extends QueryTest with SQLTestUtils {
+  import testImplicits._
 
   test("create a permanent view on a permanent view") {
     withView("jtv1", "jtv2") {
@@ -85,7 +90,7 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
 
   test("error handling: existing a table with the duplicate name when creating/altering a view") {
     withTable("tab1") {
-      sql("CREATE TABLE tab1 (id int)")
+      sql("CREATE TABLE tab1 (id int) USING parquet")
       var e = intercept[AnalysisException] {
         sql("CREATE OR REPLACE VIEW tab1 AS SELECT * FROM jt")
       }.getMessage
@@ -103,7 +108,7 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
 
   test("existing a table with the duplicate name when CREATE VIEW IF NOT EXISTS") {
     withTable("tab1") {
-      sql("CREATE TABLE tab1 (id int)")
+      sql("CREATE TABLE tab1 (id int) USING parquet")
       sql("CREATE VIEW IF NOT EXISTS tab1 AS SELECT * FROM jt")
       checkAnswer(sql("select count(*) FROM tab1"), Row(0))
     }
@@ -144,8 +149,9 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
       }.getMessage
       assert(e.contains("Inserting into an RDD-based table is not allowed"))
 
-      val testData = hiveContext.getHiveFile("data/files/employee.dat").getCanonicalPath
-      assertNoSuchTable(s"""LOAD DATA LOCAL INPATH "$testData" INTO TABLE $viewName""")
+      val dataFilePath =
+        Thread.currentThread().getContextClassLoader.getResource("data/files/employee.dat")
+      assertNoSuchTable(s"""LOAD DATA LOCAL INPATH "$dataFilePath" INTO TABLE $viewName""")
       assertNoSuchTable(s"TRUNCATE TABLE $viewName")
       assertNoSuchTable(s"SHOW CREATE TABLE $viewName")
       assertNoSuchTable(s"SHOW PARTITIONS $viewName")
@@ -169,9 +175,10 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
       }.getMessage
       assert(e.contains("Inserting into an RDD-based table is not allowed"))
 
-      val testData = hiveContext.getHiveFile("data/files/employee.dat").getCanonicalPath
+      val dataFilePath =
+        Thread.currentThread().getContextClassLoader.getResource("data/files/employee.dat")
       e = intercept[AnalysisException] {
-        sql(s"""LOAD DATA LOCAL INPATH "$testData" INTO TABLE $viewName""")
+        sql(s"""LOAD DATA LOCAL INPATH "$dataFilePath" INTO TABLE $viewName""")
       }.getMessage
       assert(e.contains(s"Target table in LOAD DATA cannot be a view: `default`.`testview`"))
 
@@ -264,16 +271,16 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
   }
 
   test("correctly parse a view with custom column names") {
-    withTable("testTable") {
-      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("testTable")
+    withTable("tab1") {
+      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("tab1")
       withView("testView1", "testView2") {
         // Correctly create a view with custom column names
-        sql("CREATE VIEW testView1(x, y) AS SELECT * FROM testTable")
+        sql("CREATE VIEW testView1(x, y) AS SELECT * FROM tab1")
         checkAnswer(sql("SELECT * FROM testView1 ORDER BY x"), (1 to 9).map(i => Row(i, i + 1)))
 
         // Throw an AnalysisException if the number of columns don't match up.
         val e = intercept[AnalysisException] {
-          sql("CREATE VIEW testView2(x, y, z) AS SELECT * FROM testTable")
+          sql("CREATE VIEW testView2(x, y, z) AS SELECT * FROM tab1")
         }.getMessage
         assert(e.contains("The number of columns produced by the SELECT clause (num: `2`) does " +
           "not match the number of column names specified by CREATE VIEW (num: `3`)."))
@@ -465,26 +472,6 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
     }
   }
 
-  test("create hive view for json table") {
-    // json table is not hive-compatible, make sure the new flag fix it.
-    withView("testView") {
-      sql("CREATE VIEW testView AS SELECT id FROM jt")
-      checkAnswer(sql("SELECT * FROM testView ORDER BY id"), (1 to 9).map(i => Row(i)))
-    }
-  }
-
-  test("create hive view for partitioned parquet table") {
-    // partitioned parquet table is not hive-compatible, make sure the new flag fix it.
-    withTable("parTable") {
-      withView("testView") {
-        val df = Seq(1 -> "a").toDF("i", "j")
-        df.write.format("parquet").partitionBy("i").saveAsTable("parTable")
-        sql("CREATE VIEW testView AS SELECT i, j FROM parTable")
-        checkAnswer(sql("SELECT * FROM testView"), Row(1, "a"))
-      }
-    }
-  }
-
   test("CTE within view") {
     withView("cte_view") {
       sql("CREATE VIEW cte_view AS WITH w AS (SELECT 1 AS n) SELECT n FROM w")
@@ -494,15 +481,15 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
 
   test("Using view after switching current database") {
     withView("v") {
-      sql("CREATE VIEW v AS SELECT * FROM src")
+      sql("CREATE VIEW v AS SELECT * FROM jt")
       withTempDatabase { db =>
         activateDatabase(db) {
-          // Should look up table `src` in database `default`.
-          checkAnswer(sql("SELECT * FROM default.v"), sql("SELECT * FROM default.src"))
+          // Should look up table `jt` in database `default`.
+          checkAnswer(sql("SELECT * FROM default.v"), sql("SELECT * FROM default.jt"))
 
-          // The new `src` table shouldn't be scanned.
-          sql("CREATE TABLE src(key INT, value STRING)")
-          checkAnswer(sql("SELECT * FROM default.v"), sql("SELECT * FROM default.src"))
+          // The new `jt` table shouldn't be scanned.
+          sql("CREATE TABLE jt(key INT, value STRING) USING parquet")
+          checkAnswer(sql("SELECT * FROM default.v"), sql("SELECT * FROM default.jt"))
         }
       }
     }
@@ -519,109 +506,13 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
     }
   }
 
-  test("create hive view for joined tables") {
-    // make sure the new flag can handle some complex cases like join and schema change.
-    withTable("jt1", "jt2") {
-      spark.range(1, 10).toDF("id1").write.format("json").saveAsTable("jt1")
-      spark.range(1, 10).toDF("id2").write.format("json").saveAsTable("jt2")
-      withView("testView") {
-        sql("CREATE VIEW testView AS SELECT * FROM jt1 JOIN jt2 ON id1 == id2")
-        checkAnswer(sql("SELECT * FROM testView ORDER BY id1"), (1 to 9).map(i => Row(i, i)))
-
-        val df = (1 until 10).map(i => i -> i).toDF("id1", "newCol")
-        df.write.format("json").mode(SaveMode.Overwrite).saveAsTable("jt1")
-        checkAnswer(sql("SELECT * FROM testView ORDER BY id1"), (1 to 9).map(i => Row(i, i)))
-      }
-    }
-  }
-
-  test("SPARK-14933 - create view from hive parquet table") {
-    withTable("t_part") {
-      withView("v_part") {
-        spark.sql("create table t_part stored as parquet as select 1 as a, 2 as b")
-        spark.sql("create view v_part as select * from t_part")
-        checkAnswer(
-          sql("select * from t_part"),
-          sql("select * from v_part"))
-      }
-    }
-  }
-
-  test("SPARK-14933 - create view from hive orc table") {
-    withTable("t_orc") {
-      withView("v_orc") {
-        spark.sql("create table t_orc stored as orc as select 1 as a, 2 as b")
-        spark.sql("create view v_orc as select * from t_orc")
-        checkAnswer(
-          sql("select * from t_orc"),
-          sql("select * from v_orc"))
-      }
-    }
-  }
-
-  test("create a permanent/temp view using a hive, built-in, and permanent user function") {
-    val permanentFuncName = "myUpper"
-    val permanentFuncClass =
-      classOf[org.apache.hadoop.hive.ql.udf.generic.GenericUDFUpper].getCanonicalName
-    val builtInFuncNameInLowerCase = "abs"
-    val builtInFuncNameInMixedCase = "aBs"
-    val hiveFuncName = "histogram_numeric"
-
-    withUserDefinedFunction(permanentFuncName -> false) {
-      sql(s"CREATE FUNCTION $permanentFuncName AS '$permanentFuncClass'")
-      withTable("tab1") {
-        (1 to 10).map(i => (s"$i", i)).toDF("str", "id").write.saveAsTable("tab1")
-        Seq("VIEW", "TEMPORARY VIEW").foreach { viewMode =>
-          withView("view1") {
-            sql(
-              s"""
-                 |CREATE $viewMode view1
-                 |AS SELECT
-                 |$permanentFuncName(str),
-                 |$builtInFuncNameInLowerCase(id),
-                 |$builtInFuncNameInMixedCase(id) as aBs,
-                 |$hiveFuncName(id, 5) over()
-                 |FROM tab1
-               """.stripMargin)
-            checkAnswer(sql("select count(*) FROM view1"), Row(10))
-          }
-        }
-      }
-    }
-  }
-
-  test("create a permanent/temp view using a temporary function") {
-    val tempFunctionName = "temp"
-    val functionClass =
-      classOf[org.apache.hadoop.hive.ql.udf.generic.GenericUDFUpper].getCanonicalName
-    withUserDefinedFunction(tempFunctionName -> true) {
-      sql(s"CREATE TEMPORARY FUNCTION $tempFunctionName AS '$functionClass'")
-      withView("view1", "tempView1") {
-        withTable("tab1") {
-          (1 to 10).map(i => s"$i").toDF("id").write.saveAsTable("tab1")
-
-          // temporary view
-          sql(s"CREATE TEMPORARY VIEW tempView1 AS SELECT $tempFunctionName(id) from tab1")
-          checkAnswer(sql("select count(*) FROM tempView1"), Row(10))
-
-          // permanent view
-          val e = intercept[AnalysisException] {
-            sql(s"CREATE VIEW view1 AS SELECT $tempFunctionName(id) from tab1")
-          }.getMessage
-          assert(e.contains("Not allowed to create a permanent view `view1` by referencing " +
-            s"a temporary function `$tempFunctionName`"))
-        }
-      }
-    }
-  }
-
   test("error handling: fail if the referenced table or view is invalid") {
     withView("view1", "view2", "view3") {
       // Fail if the referenced table is defined in a invalid database.
       withTempDatabase { db =>
         withTable(s"$db.table1") {
           activateDatabase(db) {
-            sql("CREATE TABLE table1(a int, b string)")
+            sql("CREATE TABLE table1(a int, b string) USING parquet")
             sql("CREATE VIEW default.view1 AS SELECT * FROM table1")
           }
         }
@@ -630,7 +521,7 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
 
       // Fail if the referenced table is invalid.
       withTable("table2") {
-        sql("CREATE TABLE table2(a int, b string)")
+        sql("CREATE TABLE table2(a int, b string) USING parquet")
         sql("CREATE VIEW view2 AS SELECT * FROM table2")
       }
       assertInvalidReference("SELECT * FROM view2")
@@ -644,32 +535,6 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
     }
   }
 
-  test("make sure we can resolve view created by old version of Spark") {
-    withTable("hive_table") {
-      withView("old_view") {
-        spark.sql("CREATE TABLE hive_table AS SELECT 1 AS a, 2 AS b")
-        // The views defined by older versions of Spark(before 2.2) will have empty view default
-        // database name, and all the relations referenced in the viewText will have database part
-        // defined.
-        val view = CatalogTable(
-          identifier = TableIdentifier("old_view"),
-          tableType = CatalogTableType.VIEW,
-          storage = CatalogStorageFormat.empty,
-          schema = new StructType().add("a", "int").add("b", "int"),
-          viewText = Some("SELECT `gen_attr_0` AS `a`, `gen_attr_1` AS `b` FROM (SELECT " +
-            "`gen_attr_0`, `gen_attr_1` FROM (SELECT `a` AS `gen_attr_0`, `b` AS " +
-            "`gen_attr_1` FROM hive_table) AS gen_subquery_0) AS hive_table")
-        )
-        hiveContext.sessionState.catalog.createTable(view, ignoreIfExists = false)
-        val df = sql("SELECT * FROM old_view")
-        // Check the output rows.
-        checkAnswer(df, Row(1, 2))
-        // Check the output schema.
-        assert(df.schema.sameType(view.schema))
-      }
-    }
-  }
-
   test("correctly resolve a view in a self join") {
     withView("testView") {
       sql("CREATE VIEW testView AS SELECT * FROM jt")
@@ -680,46 +545,46 @@ class SQLViewSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
   }
 
   test("resolve a view with custom column names") {
-    withTable("testTable") {
-      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("testTable")
+    withTable("tab1") {
+      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("tab1")
       withView("testView") {
-        sql("CREATE VIEW testView(x, y) AS SELECT * FROM testTable")
+        sql("CREATE VIEW testView(x, y) AS SELECT * FROM tab1")
 
         // Correctly resolve a view with custom column names.
         checkAnswer(sql("SELECT * FROM testView ORDER BY x"), (1 to 9).map(i => Row(i, i + 1)))
 
         // Correctly resolve a view when the referenced table schema changes.
         spark.range(1, 10).selectExpr("id", "id + id dummy", "id + 1 id1")
-          .write.mode(SaveMode.Overwrite).saveAsTable("testTable")
+          .write.mode(SaveMode.Overwrite).saveAsTable("tab1")
         checkAnswer(sql("SELECT * FROM testView ORDER BY x"), (1 to 9).map(i => Row(i, i + 1)))
 
         // Throw an AnalysisException if the column name is not found.
         spark.range(1, 10).selectExpr("id", "id + 1 dummy")
-          .write.mode(SaveMode.Overwrite).saveAsTable("testTable")
+          .write.mode(SaveMode.Overwrite).saveAsTable("tab1")
         intercept[AnalysisException](sql("SELECT * FROM testView"))
       }
     }
   }
 
   test("resolve a view when the dataTypes of referenced table columns changed") {
-    withTable("testTable") {
-      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("testTable")
+    withTable("tab1") {
+      spark.range(1, 10).selectExpr("id", "id + 1 id1").write.saveAsTable("tab1")
       withView("testView") {
-        sql("CREATE VIEW testView AS SELECT * FROM testTable")
+        sql("CREATE VIEW testView AS SELECT * FROM tab1")
 
         // Allow casting from IntegerType to LongType
         val df = (1 until 10).map(i => (i, i + 1)).toDF("id", "id1")
-        df.write.format("json").mode(SaveMode.Overwrite).saveAsTable("testTable")
+        df.write.format("json").mode(SaveMode.Overwrite).saveAsTable("tab1")
         checkAnswer(sql("SELECT * FROM testView ORDER BY id1"), (1 to 9).map(i => Row(i, i + 1)))
 
         // Casting from DoubleType to LongType might truncate, throw an AnalysisException.
         val df2 = (1 until 10).map(i => (i.toDouble, i.toDouble)).toDF("id", "id1")
-        df2.write.format("json").mode(SaveMode.Overwrite).saveAsTable("testTable")
+        df2.write.format("json").mode(SaveMode.Overwrite).saveAsTable("tab1")
         intercept[AnalysisException](sql("SELECT * FROM testView"))
 
         // Can't cast from ArrayType to LongType, throw an AnalysisException.
         val df3 = (1 until 10).map(i => (i, Seq(i))).toDF("id", "id1")
-        df3.write.format("json").mode(SaveMode.Overwrite).saveAsTable("testTable")
+        df3.write.format("json").mode(SaveMode.Overwrite).saveAsTable("tab1")
         intercept[AnalysisException](sql("SELECT * FROM testView"))
       }
     }
